redisson을 사용한 동시성 처리

+++ src/main/java/kb/kiomnd2/kbblogsearch/jpa/domain/SearchEntity.java
@@ -22,8 +22,10 @@ public class SearchEntity {
 
     private LocalDateTime createAt;
 
+/*
     @Version
-    private Long version;
+    private int version;
+*/
 
     @Builder
     public SearchEntity(String keyword, long count, LocalDateTime createAt) {
